requirements.txt
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0
tensorflow>=2.15.0
keras>=2.15.0
matplotlib>=3.8.0
seaborn>=0.13.0
flask>=2.3.0
flask-cors>=4.0.0
psutil>=5.9.0
scapy>=2.5.0
plotly>=5.18.0
dash>=2.14.0
dash-bootstrap-components>=1.5.0
joblib>=1.3.0
python-dotenv>=1.0.0
schedule>=1.2.0

